// src/app/enums/apiResponse.ts
export interface WebMemberTypeModel {
  statuscode: number;
  msg: string;
  memberTypes: MembershipmasterB2C[];
}
export interface MembershipmasterB2C {
  isIDActive: boolean;
  id: number;
  memberShipType: string;
  couponCount: number;
  couponValue: number;
  isCouponAllowed: boolean;
  remark: string;
  couponValidityDays: number;
  cost: number;
}
// ...
export interface CouponDetail {
  id: number;
  couponCode: string;
  couponExpiry: string;
  redeemDate: string;
  isRedeemed: boolean;
  couponValue: number;
}
//#region  Transection
export interface TransectionResp {
  liveID: string;
  transactionID: string;
  statuscode: number;
  msg: string;
  isVersionValid: boolean;
  isAppValid: boolean;
  checkID: number;
  isPasswordExpired: boolean;
  mobileNo?: any;
  emailID?: any;
  isLookUpFromAPI: boolean;
  isDTHInfoCall: boolean;
  isShowPDFPlan: boolean;
  sid?: any;
  isOTPRequired: boolean;
  getID: number;
}